add register of policies on athServuceProvider

+++ src/Generators/ControllerGenerator.php
@@ -54,8 +54,9 @@ class {$this->modelName}Controller extends Controller
      */
     public function index(Request \$request): AnonymousResourceCollection
     {
+        // Policy: viewAny
+        $this->authorize('viewAny', {$this->modelName}::class);
         \$result = \$this->service->search(\$request->all());
-
         return {$this->modelName}Resource::collection(\$result);
     }
 
@@ -64,8 +65,9 @@ class {$this->modelName}Controller extends Controller
      */
     public function store(Store{$this->modelName}Request \$request): {$this->modelName}Resource
     {
+        // Policy: create
+        $this->authorize('create', {$this->modelName}::class);
         \${$this->modelNameCamel} = \$this->service->create(\$request->validated());
-
         return new {$this->modelName}Resource(\${$this->modelNameCamel});
     }
 
@@ -74,6 +76,8 @@ class {$this->modelName}Controller extends Controller
      */
     public function show({$this->modelName} \${$this->modelNameCamel}): {$this->modelName}Resource
     {
+        // Policy: view
+        $this->authorize('view', \${$this->modelNameCamel});
         return new {$this->modelName}Resource(\${$this->modelNameCamel});
     }
 
@@ -84,11 +88,12 @@ class {$this->modelName}Controller extends Controller
         Update{$this->modelName}Request \$request,
         {$this->modelName} \${$this->modelNameCamel}
     ): {$this->modelName}Resource {
+        // Policy: update
+        $this->authorize('update', \${$this->modelNameCamel});
         \${$this->modelNameCamel} = \$this->service->update(
             \${$this->modelNameCamel},
             \$request->validated()
         );
-
         return new {$this->modelName}Resource(\${$this->modelNameCamel});
     }
 
@@ -97,8 +102,9 @@ class {$this->modelName}Controller extends Controller
      */
     public function destroy({$this->modelName} \${$this->modelNameCamel}): JsonResponse
     {
+        // Policy: delete
+        $this->authorize('delete', \${$this->modelNameCamel});
         \$this->service->delete(\${$this->modelNameCamel});
-
         return response()->json(null, 204);
     }
 }
